requirements.txt
requests
Pillow
numpy
boto3
transformers>=4.49.0
torch
torchvision
mediapipe
qwen-vl-utils
git+https://github.com/IDEA-Research/GroundingDINO.git